Update thalamus (#18)
* Add support for kernel radius in simple_blur
* Add thalamus dirvec changes
* Run isort and black as in readme
* Revert RT direction
* update copyright
* Modify thalamus dirvecs and tests so it passes

This makes a small change to the target of the "RT-complement" part of
the thalamus direction-vectors calculation. This change prevents
thalamic NaN direction-vectors from being created in the test case. It
should be noted that, when applied to the "real data" Atlas annotation
instead of the test case, neither of these landscapes produced any NaN
direction vectors, despite the previous landscape producing NaNs in the
test case. The new landscape, when applied to Atlas data, seems to
produce direction-vectors which are satisfactory (if interested, ask me
for details).

In order to pass the tests, I also had to reduce the tolerance on one of
the norm tests, since in the test case, it was producing norms that
further off from 1 than 1e-6. If such norms are produced in the actual
data, I'm not sure what the effects will be on placement-hints, but my
run of placement-hints using real data from these new direction-vectors
seemd to be fine. Also, I will point out that when running the test case
for the pre-2023 values of the direction vectors (including `sigma=ratio
* 18.0` etc.), the vast majority of the test case voxels were not
changed from their initialized values (i.e., they were not within
range of the gaussian blur being applied at all). This may have led to
most voxels passing this norm test as a false positive, since most
voxels in the older test case were not actually modified at all, if
being affected by the gaussian blur (meant to point them to RT, which
they mostly were not in this case) is what likely has the capacity to
change the norm from 1. In other words: the test whose tolerance I
changed was being passed in the most recent test case, but I suspect
that this is because the direction-vector algorithm was not changing the
vast majority of voxels in the test case, when it should have been.