Limit thalamus region dirvecs to specific regions
This replaces the region selection of what region direction-vectors are
made for in thalamus with a regular expression. This was scientifically
necessary for exclusion of the habenular and peripeduncular regions, and
code-wise so that the thalamus region selection regex can be harmonized
across all uses. For information on which regions were chosen and how
this list was created, see the internal BBP Confluence page located at
"Circuits > Mouse Thalamus > Atlas-based Whole-thalamus subregion
selection". This regex has been built from the region list of the
desired and present thalamus regions as of the "final" version of the
hierarchy and annotation built by the Atlas pipeline, which is the
output of the rule `split_barrel_ccfv3_l23split`.